cell bug fix 2
